Refactor event handling in Bar chart.
